Add default end date/time for events with no end time or duration
Before this change, some events would be parsed and result in a
ValueError. According to RFC 5545 section 3.6.1, there are default
behaviors to be assumed in these cases - namely that events with a
starting date should assume the event is one full day, while events
with a starting time and date should assume zero duration.