Fix confusing message when TZID not recognized -- Full Version (#499)
Fix confusing message when TZID not recognized

When validating lines of an ICS file, we may not know the data-type of
the property value: perhaps it is a 20250601 date or a 20250601T171819
date-time. The validator tries each potential data-type, and uses the
first successfully validated type. What happens if all report invalid?

We now distinguish two types of validation errors:
- ValueError: the property value is not of this type
- ParameterValueError: it is of this type but is not valid

A ParameterValueError will stop the checking and report only that error;
otherwise all ValueErrors are reported.

If a ParameterValueError is encountered, for example in:
  DTSTART;TZID=America/New_Mork:20250601T171819
then the new error message becomes:
  Failed to parse calendar EVENT component:
  Expected DATE-TIME TZID value 'America/New_Mork' to be valid timezone
and no longer mentions that it is also not a valid date.